\name{10.Other}
\alias{10.Other}
\title{Other Functions}

\description{
This page describes some functions not covered in the previous numbered pages, so far only \code{\link{blockDiag}} and \code{\link{poolVar}} which are not used in the package yet but are part of the development of methods to handle technical and biological replicates.
}

\author{Gordon Smyth}
\keyword{documentation}
