\name{1.Introduction}
\alias{1.Introduction}
\title{Introduction to the LIMMA Package}
\description{
LIMMA is a library for the analysis of gene expression microarray data, especially the use of linear models for analysing designed experiments and the assessment of differential expression.
The package is mostly for two-colour microarrays but the linear model and differential expression routines are equally applicable to multi-array oligonucleotide experiments.

See "Accompanying documentation" at the top of the LIMMA index page for the LIMMA User's Guide.
}
\author{Gordon Smyth}
\references{
Smyth, G. K., Yang, Y.-H., Speed, T. P. (2003). Statistical issues in microarray data analysis. In: \emph{Functional Genomics: Methods and Protocols}, M. J. Brownstein and A. B. Khodursky (eds.), Methods in Molecular Biology Volume 224, Humana Press, Totowa, NJ, pages 111-136.
}
\keyword{documentation}
