\name{limmaguideURL}
\alias{limmaguideURL}

\title{File Location of html Limma User's Guide}
\description{Finds the location of the html version of the Limma User's Guide.}

\usage{
limmaguideURL()
}

\value{Character string giving the file location as a URL.}

\details{
This function is used by the Window drop-down Vignettes menu.
}

\author{Gordon Smyth}

\keyword{documentation}
