\name{01.Introduction}
\alias{01.Introduction}
\alias{limma}
\title{Introduction to the LIMMA Package}
\description{
LIMMA is a library for the analysis of gene expression microarray data, especially the use of linear models for analysing designed experiments and the assessment of differential expression.
LIMMA provides the ability to analyse comparisons between many RNA targets simultaneously in arbitrary complicated designed experiments.
Empirical Bayesian methods are used to provide stable results even when the number of arrays is small.
The normalization and data analysis functions are for two-colour spotted microarrays.
The linear model and differential expression functions apply to all microarrays including Affymetrix and other multi-array oligonucleotide experiments.

There are three types of documentation available.
(1) The \emph{LIMMA User's Guide} can be reached through the "User Guides and Package Vignettes" links at the top of the LIMMA contents page.
The function \code{\link{limmaUsersGuide}} gives the file location of the User's Guide.
(2) An overview of limma functions grouped by purpose is contained in the numbered chapters at the top of the \link[limma:00Index]{LIMMA contents page}, of which this page is the first.
(3) The \link[limma:00Index]{LIMMA contents page} gives an alphabetical index of detailed help topics.

The function \code{\link{changeLog}} displays the record of changes to the package.
}
\author{Gordon Smyth}
\references{
Smyth, G. K., Yang, Y.-H., Speed, T. P. (2003). Statistical issues in microarray data analysis. \emph{Methods in Molecular Biology} 224, 111-136.

Smyth, G. K. (2005). Limma: linear models for microarray data. In: 'Bioinformatics and Computational Biology Solutions using R and Bioconductor'. R. Gentleman, V.   Carey, S. Dudoit, R. Irizarry, W. Huber (eds), Springer, New York, 2005.
}
\keyword{documentation}
