\name{1.Introduction}
\alias{1.Introduction}
\title{Introduction to the LIMMA Package}
\description{
LIMMA is a library for the analysis of gene expression microarray data, especially the use of linear models for analysing designed experiments and the assessment of differential expression.
LIMMA provides the ability to analyse comparisons between many RNA targets simultaneously.
The normalization and data analysis functions are for two-colour spotted microarrays.
The linear model and differential expression functions apply to all microarrays including Affymetrix and other multi-array oligonucleotide experiments.

There are three types of documentation available:
\describe{
\item{\emph{LIMMA User's Guide}}{
	See "Accompanying documentation" at the top of the LIMMA contents page.}
\item{Alphabetical index of functions and classes}{
	See the LIMMA contents page}
\item{Grouping of functions by purpose}{
	See the numbered chapters at the top of the LIMMA contents page.}
}
}
\author{Gordon Smyth}
\references{
Smyth, G. K., Yang, Y.-H., Speed, T. P. (2003). Statistical issues in microarray data analysis. In: \emph{Functional Genomics: Methods and Protocols}, M. J. Brownstein and A. B. Khodursky (eds.), Methods in Molecular Biology Volume 224, Humana Press, Totowa, NJ, pages 111-136.
}
\keyword{documentation}
