\name{as.MAList}
\alias{as.MAList}
\title{Convert marrayNorm Object to an MAList Object}

\description{
Convert marrayNorm Object to an MAList Object
}

\usage{
as.MAList(object)
}

\arguments{
\item{object}{an \code{\link[marrayClasses:marrayNorm-class]{marrayNorm}} object}
}

\value{
Object of class \code{\link[limma:MAList]{MAList}}
}

\author{Gordon Smyth}

\seealso{
  \link{02.Classes} gives an overview of all the classes defined by this package.
}

\keyword{classes}
\keyword{data}
