\name{changeLog}
\alias{changeLog}

\title{Limma Change Log}
\description{Write as text the most recent changes from the limma package changelog.}

\usage{
changeLog(n=20)
}

\arguments{
  \item{n}{integer, number of lines to write of changelog.}
}

\value{No value is produced, but a number of lines of text are written to standard output.}

\author{Gordon Smyth}

\seealso{
\link{01.Introduction}
}

\keyword{documentation}
