\name{fitted.MArrayLM}
\alias{fitted.MArrayLM}
\concept{regression}
\title{Fitted Values Method for MArrayLM Fits}
\usage{
\S3method{fitted}{MArrayLM}(object, \dots)
}
\arguments{
  \item{object}{a fitted object of class inheriting from \code{"MArrayLM"}.}
  \item{\dots}{other arguments are not currently used.}
}
\description{
  Obtains fitted values from a fitted microarray linear model object.
}
\value{
  A numeric matrix of fitted values.
}
\seealso{
  \code{\link{fitted}}
}

\author{Gordon Smyth}
\keyword{models}
\keyword{regression}
