\name{m.spot}
\alias{m.spot}
\alias{a.spot}
\title{Extract M or A-values from SPOT data.frame or matrix}
\description{
Extract M-values or A-values from a SPOT data.frame or matrix.

These functions are deprecated in limma 2.9.7, 1 January 2007.
}
\usage{
m.spot(spot)
a.spot(spot)
}
\arguments{
  \item{spot}{data.frame or matrix giving SPOT output for one microarray}
}
\value{
Vector of M-values (\code{m.spot}) or A-values (\code{a.spot})
}
\author{Gordon Smyth}
\keyword{IO}
